AND-8562 Move TestnetTokensStorage to local.config package
