+++ core/datasource/src/main/java/com/tangem/datasource/local/config/testnet/models/TestnetTokensConfig.kt
@@ -1,4 +1,4 @@
-package com.tangem.datasource.local.testnet.models
+package com.tangem.datasource.local.config.testnet.models
 
 import com.squareup.moshi.Json
 
@@ -7,6 +7,7 @@ import com.squareup.moshi.Json
  *
  * @property tokens testnet tokens list
  */
+// TODO: use CoinsResponse
 data class TestnetTokensConfig(
     @Json(name = "coins") val tokens: List<Token>,
 ) {
